完成Mapper.xml类的查询功能，并测试了查询功能的可行性

+++ easyJavaDemo/src/main/java/com/easyjava/mappers/BaseMapper.java
@@ -6,19 +6,19 @@ public interface BaseMapper<T,P>{
     /*
     insert:插入
      */
-    Integer insert(@Param("bean") T t);
+    Long insert(@Param("bean") T t);
     /*
     插入或者更新
      */
-    Integer insertOrUpdate(@Param("bean") T t);
+    Long insertOrUpdate(@Param("bean") T t);
     /*
     批量插入
      */
-    Integer insertBatch(@Param("list") List<T> list);
+    Long insertBatch(@Param("list") List<T> list);
     /*
     批量插入或者更新
      */
-    Integer insertOrUpdateBatch(@Param("list") List<T> list);
+    Long insertOrUpdateBatch(@Param("list") List<T> list);
     /*
     根据参数查询集合
      */
@@ -26,5 +26,5 @@ public interface BaseMapper<T,P>{
     /*
     根据集合查询数量
      */
-    Integer selectCount(@Param("query") P p);
+    Long selectCount(@Param("query") P p);
 }
